All sound effects are available now
I added the sound effects that we need and get them tested, please check them out!

I noticed that the version you uploaded on github is quite different to the version you showed me on Wednesday, for the ball is bouncing kinda weird. I think we can work on that next.