Levels design complete
All levels are now functioning well!

Level 1: the ball destroy the block everytime it hits it, and gain scores;

Level 2: the ball will let the block drop everytime it hits it, player have to collect the blocks to gain scores;

Level 3: same as level 2, but now the player has four balls to juggle with.

Art materials fixed.